Address Claude review on PR 1298: elevate inline GQL, harden test wrappers
- Move GET_CORPUS_AGENTS / GET_AVAILABLE_TOOLS from inline defs in
  CorpusAgentManagement.tsx into src/graphql/queries.ts, and move the three
  agent-config mutations into src/graphql/mutations.ts. Both the production
  component and CorpusAgentManagement.ct.tsx now import from the shared
  modules so test mocks can never drift from the query actually executed.
- Add explicit InMemoryCache to CorpusAgentManagementTestWrapper,
  CorpusDescriptionEditorTestWrapper, and CreateCorpusActionModalTestWrapper
  to match the CLAUDE.md pitfall #8 pattern already used by
  CorpusChatTestWrapper.
- Move authToken/userObj priming from the CorpusChatTestWrapper render body
  to module scope so the reactive vars are seeded exactly once before any
  render, avoiding a side effect on every re-render (StrictMode safe).
- Log JSON parse errors from StubSocket.send() in CorpusChat.ct.tsx instead
  of swallowing them silently — a timeout here now gets a clear warning
  rather than a mystery hang.
