+++ frontend/tests/CorpusAgentManagementTestWrapper.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { MockedProvider, MockedResponse } from "@apollo/client/testing";
+import { InMemoryCache } from "@apollo/client";
 import { Provider as JotaiProvider } from "jotai";
 import { ToastContainer } from "react-toastify";
 import { CorpusAgentManagement } from "../src/components/corpuses/CorpusAgentManagement";
@@ -21,9 +22,12 @@ export const CorpusAgentManagementTestWrapper: React.FC<Props> = ({
   corpusId,
   canUpdate = true,
 }) => {
+  // Defined inside the wrapper so Playwright CT's per-test serialization
+  // never reaches an Apollo cache instance — see CLAUDE.md pitfall #8.
+  const cache = new InMemoryCache({ addTypename: false });
   return (
     <JotaiProvider>
-      <MockedProvider mocks={mocks} addTypename={false}>
+      <MockedProvider mocks={mocks} addTypename={false} cache={cache}>
         <div style={{ width: "100vw", padding: 16 }}>
           <CorpusAgentManagement corpusId={corpusId} canUpdate={canUpdate} />
           <ToastContainer />
